requirements.txt
tensorflow-cpu
absl-py==0.8.1
asgiref==3.2.3
asn1crypto==1.3.0
astor==0.8.0
blinker==1.4
cachetools==3.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
imageio==2.6.1
itsdangerous==1.1.0
Jinja2==2.11.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
music21==5.7.2
numpy==1.16.6
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
pandas==1.0.1
Pillow==7.0.0
protobuf==3.11.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
tensorboard==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
Theano==1.0.4
tornado==6.0.3
urllib3==1.25.8
whitenoise==5.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
